Changes for Justin & Gary comments.
Fix index bug in manipulating expected groups for dist.
