+++ pipelines/disturbance/prefect_flows/gadm_dist_alerts.py
@@ -5,6 +5,7 @@ from prefect import flow
 from pipelines.disturbance.prefect_flows import dist_common_tasks
 from pipelines.globals import DATA_LAKE_BUCKET
 from pipelines.utils import s3_uri_exists
+from pipelines.prefect_flows import common_tasks
 
 @flow(name="DIST alerts count")
 def dist_alerts_count(dist_zarr_uri: str, dist_version: str, overwrite=False):
@@ -20,20 +21,20 @@ def dist_alerts_count(dist_zarr_uri: str, dist_version: str, overwrite=False):
         np.arange(731, 1590),  # dates values
         [1, 2, 3],  # confidence values
     )
-    datasets = dist_common_tasks.load_data.with_options(name="dist-alerts-load_data")(
+    datasets = common_tasks.load_data.with_options(name="dist-alerts-load_data")(
         dist_zarr_uri
     )
     compute_input = dist_common_tasks.setup_compute.with_options(
         name="set-up-dist-alerts-compute"
     )(datasets, expected_groups)
 
-    result_dataset = dist_common_tasks.compute_zonal_stat.with_options(
+    result_dataset = common_tasks.compute_zonal_stat.with_options(
         name="dist-alerts-compute-zonal-stats"
     )(*compute_input, funcname="count")
-    result_df = dist_common_tasks.postprocess_result.with_options(
+    result_df = common_tasks.postprocess_result.with_options(
         name="dist-alerts-postprocess-result"
     )(result_dataset)
-    dist_common_tasks.save_result.with_options(
+    common_tasks.save_result.with_options(
         name="dist-alerts-save-result"
     )(result_df, result_uri)
