+++ pipelines/natural_lands/prefect_flows/nl_flow.py
@@ -7,6 +7,7 @@ from pipelines.globals import DATA_LAKE_BUCKET
 from pipelines.utils import s3_uri_exists
 from pipelines.disturbance.prefect_flows import dist_common_tasks
 from pipelines.natural_lands.prefect_flows import nl_tasks
+from pipelines.prefect_flows import common_tasks
 
 
 @flow(name="Natural lands area")
@@ -18,7 +19,7 @@ def gadm_natural_lands_area(
     base_uri = "s3://gfw-data-lake/umd_area_2013/v1.10/raster/epsg-4326/zarr/pixel_area.zarr"
     contextual_uri = "s3://gfw-data-lake/sbtn_natural_lands/zarr/sbtn_natural_lands_all_classes.zarr"
     contextual_column_name = "natural_lands"
-    result_uri = f's3://{DATA_LAKE_BUCKET}/sbtn_natural_lands/zarr/area_by_natural_lands_all_adm2_raw.parquet'
+    result_uri = f's3://{DATA_LAKE_BUCKET}/sbtn_natural_lands/tabular/zonal_stats/gadm/gadm_adm2.parquet'
     funcname = "sum"
 
     if not overwrite and s3_uri_exists(result_uri):
@@ -31,7 +32,7 @@ def gadm_natural_lands_area(
         np.arange(1, 22),   # natural lands categories
     )
 
-    datasets = dist_common_tasks.load_data.with_options(
+    datasets = common_tasks.load_data.with_options(
         name="area-by-natural-lands-load-data"
     )(base_uri, contextual_uri=contextual_uri)
 
@@ -39,13 +40,13 @@ def gadm_natural_lands_area(
         name="set-up-area-by-natural-lands-compute"
     )(datasets, expected_groups, contextual_name=contextual_column_name)
 
-    result_dataset = dist_common_tasks.compute_zonal_stat.with_options(
+    result_dataset = common_tasks.compute_zonal_stat.with_options(
         name="area-by-natural-lands-compute-zonal-stats"
     )(*compute_input, funcname=funcname)
-    result_df = dist_common_tasks.postprocess_result.with_options(
+    result_df = common_tasks.postprocess_result.with_options(
         name="area-by-natural-lands-postprocess-result"
     )(result_dataset)
-    result_uri = dist_common_tasks.save_result.with_options(
+    result_uri = common_tasks.save_result.with_options(
         name="area-by-natural-lands-save-result"
     )(result_df, result_uri)
     return result_uri
