+++ test/pipelines/disturbance/test_gadm_dist_alerts_by_driver.py
@@ -9,8 +9,8 @@ from pipelines.disturbance.prefect_flows import dist_alerts_by_drivers_count
 
 @pytest.mark.integration
 @pytest.mark.slow
-@patch("pipelines.disturbance.stages._save_parquet")
-@patch("pipelines.disturbance.stages._load_zarr")
+@patch("pipelines.prefect_flows.common_stages._save_parquet")
+@patch("pipelines.prefect_flows.common_stages._load_zarr")
 def test_gadm_dist_alerts_by_driver_happy_path(
     mock_load_zarr,
     mock_save_parquet,
@@ -44,8 +44,8 @@ def test_gadm_dist_alerts_by_driver_happy_path(
 
 @pytest.mark.slow
 @pytest.mark.integration
-@patch("pipelines.disturbance.stages._save_parquet")
-@patch("pipelines.disturbance.stages._load_zarr")
+@patch("pipelines.prefect_flows.common_stages._save_parquet")
+@patch("pipelines.prefect_flows.common_stages._load_zarr")
 def test_gadm_dist_alerts_by_driver_result(
     mock_load_zarr,
     mock_save_parquet,
